\name{BDgraph-internal}
\alias{BDgraph-internal}

\alias{ratesf}
\alias{Exp.MC}
\alias{Psi}
\alias{sampleK}
\alias{block.gibbs.high}
\alias{block.gibbs.low}
\alias{roc}

\docType{data}

\title{ Internal bdgraph functions and datasets }

\description{ Nothing to see here move along }

\author{ Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit }

\keyword{ internal }
