\name{BDgraph-package}
\alias{BDgraph-package}
\alias{BDgraph}
\docType{package}
\title{
Bayesian model selection in Gaussian graphical models based on BDMCMC algorithm
}
\description{
BDgraph package is a statistical tool for Bayesian model determination in 
Gaussian graphical models based on birth-death MCMC methodology.      
}
\details{
\tabular{ll}{
Package: \tab BDgraph\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-10-08\cr
License: \tab GPL(>=3)\cr
}
This package provides an implementation of the procedures described in Mohammadi and Wit (2012). 
The package includes several functions. The main function is 'bdmcmc' which is 
brith-death MCMC algorithm for Bayesian model selection in Gaussian graphical models.
}
\author{
Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit
}
\references{
Mohammadi, A. and E. Wit (2012). Efficient birth-death MCMC inference for 
Gaussian graphical models, Journal of the Royal Statistical Society: Series B,
submitted.

Atay-Kayis, A. and H. Massam (2005). A monte carlo method for computing the 
marginal likelihood in nondecomposable gaussian graphical models. Biometrika 
92(2), 317-335.
}
\keyword{ package }
