\name{CellSignal}
\alias{CellSignal}
\docType{data}

\title{CellSignal dataset}

\description{
Cell signalling is a dataset with n=11673 and p = 11 proteins from Sachs et al. (2005).
}

\usage{data(CellSignal)}

\format{
  An object of class matrix.
}

\references{
  Sachs, K., O. Perez, D. Pe'er, D. Lauffenburger, and G. Nolan (2005). Causal protein-signaling networks derived from multiparameter single-cell data. Science's STKE 308 (5721), 523.
}

\keyword{datasets}
