\name{FIV}
\alias{FIV}
\docType{data}

\title{Fisher's Iris virginica dataset}

\description{ 
  Data from an experiment on for Fisher's Iris virginica dataset. This dataset 
  has been analyzed in Whittaker (1990), Giudici and Green(1999), Roverato (2002) 
  and Atay-Kayis and Massam (2005). The data gives the measurements in millimeters 
  of the sepal length, sepal width, petal length and petal width of n = 50 
  owers. We only put the symmetric matrix S = x'x for this data.
} 

\usage{data(FIV)}

\format{
  An object of class matrix. 
}  

\references{
  see \link{datasets}{iris}.
}

\keyword{datasets}
