\name{BDgraph-internal}
\alias{BDgraph-internal}

\alias{log.Exp.MC}
\alias{roc}
\alias{outRoc}

\docType{data}

\title{ Internal bdgraph functions and datasets }

\description{ Nothing to see here move along }

\author{ Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit }

\keyword{ internal }
