\name{BDgraph-internal}
\alias{BDgraph-internal}
\alias{bdgraph.high}
\alias{bdgraph.low}
\alias{Exp.MC}
\alias{Psi}
\alias{get.S}
\alias{get.cs_c}
\alias{sampleK}
\alias{block.gibbs}
\alias{roc}
\alias{CellSignal}
\docType{data}

\title{
Inernal functions and datasets
}
\description{
Nothing to see here move along
}
\author{
Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit
}
\keyword{ internal }
