\name{BDgraph-internal}
\alias{BDgraph-internal}
\alias{rates.high}
\alias{rates.low}
\alias{Exp.MC}
\alias{Psi}
\alias{get.S}
\alias{get.cs_c}
\alias{sampleK}
\alias{block.gibbs}
\alias{roc}
\docType{data}

\title{ Inernal bdgraph functions and datasets }

\description{ Nothing to see here move along }

\author{ Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit }

\keyword{ internal }
