\name{BDgraph-internal}
\alias{BDgraph-internal}
\alias{Exp.MC}
\alias{Hmatrix}
\alias{Psi}
\alias{get.S}
\alias{sampleK}

\title{
Inernal functions
}
\description{
Nothing to see here move along
}
\author{
Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit
}
\keyword{ internal }
