\name{  BDgraph-internal }
\alias{ BDgraph-internal }

\alias{ compute_tp_fp }
\alias{ compute_measures }

\alias{ hill_climb_mpl }
\alias{ local_mb_hc }
\alias{ global_hc }
\alias{ log_mpl_disrete }

\alias{ hill_climb_mpl_binary }
\alias{ local_mb_hc_binary }
\alias{ global_hc_binary }
\alias{ log_mpl_binary }

\alias{ detect_cores }

\docType{data}

\title{ Internal BDgraph functions and datasets }

\description{ Nothing to see here move along }

\author{ Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit }

\keyword{ internal }
