\name{BDgraph-internal}
\alias{BDgraph-internal}
\alias{Exp.MC}
\alias{Hmatrix}
\alias{Psi}
\alias{plotSum}
\alias{get.S}
\alias{melt}
\alias{sampleK}
\alias{CellSignal}
\alias{FIV}
\docType{data}

\title{
Inernal functions and datasets
}
\description{
Nothing to see here move along
}
\author{
Abdolreza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit
}
\keyword{ internal }
